Refactor Model and its usage Flow
Logic was tightly coupled and confusing on Pen and its dependencies.
This refactor has the goal of making the dependencies more clear and
easier to use and reuse with a simpler API.

Previously a figure was created on onPress and had a strong dependecy
on grid, now that figure is completly independent from grid until
onPressUp. This change will better reflect actions to use and send,
allowing for a simpler future implementation of websockets.