Remove dead code and other small touches
_operations shouldn't really change so there is no need for an add and
remove method. Since StringWebSocketsOperations already had a
constructor that took _operations, that should be used instead. This
also allows those operations to be private. Further discussion should
also be done about usage of ImmutableCollections in this case.

Added some TODO to check in code review. One of them is about doing the
same for the rest of the fields (Session and ClientId).

Also operations could be it's own service that could be instantiated as
less as possible to avoid creation of a "different equal" _operations
Dictionary for each request.

It could also be interesting to make a small lib to route websockets
operations using annotations (to be similar to ASP.NET Core) instead of
the Dictionary.